refactor: consolidate trip types
